# rimway-backend/docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/admin/captains": {
      "get": {
        "operationId": "AdminCaptainsController_getCaptains",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminCaptains"
        ]
      },
      "post": {
        "operationId": "AdminCaptainsController_createCaptain",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCaptainDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "AdminCaptains"
        ]
      }
    },
    "/api/v1/admin/captains/{id}": {
      "get": {
        "operationId": "AdminCaptainsController_getCaptainDetails",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "AdminCaptains"
        ]
      },
      "patch": {
        "operationId": "AdminCaptainsController_updateCaptain",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCaptainDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminCaptains"
        ]
      }
    },
    "/api/v1/admin/captains/{id}/approve": {
      "post": {
        "operationId": "AdminCaptainsController_approveCaptain",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionReasonDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "AdminCaptains"
        ]
      }
    },
    "/api/v1/admin/captains/{id}/reject": {
      "post": {
        "operationId": "AdminCaptainsController_rejectCaptain",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionReasonDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "AdminCaptains"
        ]
      }
    },
    "/api/v1/admin/captains/{id}/suspend": {
      "post": {
        "operationId": "AdminCaptainsController_suspendCaptain",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionReasonDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "AdminCaptains"
        ]
      }
    },
    "/api/v1/admin/captains/{id}/reactivate": {
      "post": {
        "operationId": "AdminCaptainsController_reactivateCaptain",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionReasonDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "AdminCaptains"
        ]
      }
    },
    "/api/v1/admin/captains/{id}/priority": {
      "post": {
        "operationId": "AdminCaptainsController_updatePriorityTier",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "AdminCaptains"
        ]
      }
    },
    "/api/v1/admin/captains/{id}/vehicle": {
      "put": {
        "operationId": "AdminCaptainsController_updateVehicle",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VehicleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "AdminCaptains"
        ]
      }
    },
    "/api/v1/admin/captains/{id}/wallet": {
      "get": {
        "operationId": "AdminCaptainsController_getWallet",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminCaptains"
        ]
      }
    },
    "/api/v1/admin/captains/{id}/wallet/transactions": {
      "get": {
        "operationId": "AdminCaptainsController_getWalletTransactions",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminCaptains"
        ]
      }
    },
    "/api/v1/admin/captains/{id}/complaints": {
      "get": {
        "operationId": "AdminCaptainsController_getComplaints",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminCaptains"
        ]
      }
    },
    "/api/v1/admin/documents/drivers/{driverId}": {
      "get": {
        "operationId": "AdminDocumentsController_getDocuments",
        "parameters": [
          {
            "name": "driverId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminDocuments"
        ]
      }
    },
...
    "/api/v1/admin/documents/{id}/status": {
      "put": {
        "operationId": "AdminDocumentsController_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminDocuments"
        ]
      }
    },
    "/api/v1/admin/ratings/drivers/{driverId}": {
      "get": {
        "operationId": "AdminRatingsController_getDriverRatings",
        "parameters": [
          {
            "name": "driverId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "AdminRatings"
        ]
      }
    },
    "/api/v1/admin/ratings": {
      "get": {
        "operationId": "AdminRatingsController_getAllRatings",
        "parameters": [
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "take",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "AdminRatings"
        ]
      }
    },
...
    "/api/v1/admin/complaints": {
      "get": {
        "operationId": "AdminComplaintsController_getComplaints",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "AdminComplaints"
        ]
      }
    },
    "/api/v1/admin/complaints/{id}/status": {
      "put": {
        "operationId": "AdminComplaintsController_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminComplaints"
        ]
      }
    },
...
    "/api/v1/admin/pricing/service-types/{id}": {
      "put": {
        "operationId": "AdminPricingController_updatePricing",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminPricing"
        ]
      }
    },
    "/api/v1/admin/audit": {
      "get": {
        "operationId": "AdminAuditController_getLogs",
        "parameters": [
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "take",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminAudit"
        ]
      }
    },
...
    "/api/v1/admin/wallet/captains/{id}": {
      "get": {
        "operationId": "AdminWalletController_getCaptainWallet",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminWallet"
        ]
      }
    },
...
  },
  "info": {
    "title": "RIM WAY API",
    "description": "The RIM WAY Backend API Contract for Passenger and Driver Apps",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "VehicleDto": {
        "type": "object",
        "properties": {
          "brand": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "year": {
            "type": "number"
          },
          "color": {
            "type": "string"
          },
          "plateNumber": {
            "type": "string"
          },
          "serviceTypeId": {
            "type": "string"
          }
        },
        "required": [
          "brand",
          "model",
          "year",
          "color",
          "plateNumber",
          "serviceTypeId"
        ]
      },
      "AdminDocumentInputDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "fileUrl": {
            "type": "string"
          },
          "fileName": {
            "type": "string"
          },
          "mimeType": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "fileUrl",
          "fileName",
          "mimeType"
        ]
      },
      "CreateCaptainDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "pattern": "^\\+222[2-4][0-9]{7}$"
          },
          "name": {
            "type": "string"
          },
          "vehicle": {
            "$ref": "#/components/schemas/VehicleDto"
          },
          "documents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminDocumentInputDto"
            }
          },
          "vehiclePhotos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminDocumentInputDto"
            }
          }
        },
        "required": [
          "phone"
        ]
      },
      "UpdateCaptainDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "photo": {
            "type": "string"
          },
          "language": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          }
        }
      },
      "ActionReasonDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string"
          }
        }
      }
    }
  }
}